feat: Setup comprehensive pre-commit hooks and code quality tools
✅ COMPLETE: Production-ready pre-commit hooks and quality tools implemented!

## What Was Implemented

### Pre-commit Hooks Configuration
- **Code formatting**: Black, isort, ruff format
- **Code quality**: Ruff linting with comprehensive rule set
- **Type checking**: MyPy with balanced strictness for development velocity
- **Security scanning**: Bandit (medium+ severity issues)
- **File validation**: JSON, YAML, TOML validation and formatting
- **General checks**: Large files, merge conflicts, trailing whitespace, etc.

### Quality Check Scripts
- `./scripts/format.sh` - Run all code formatters (Black, isort, ruff format)
- `./scripts/lint.sh` - Run linting and security checks (ruff, bandit)
- `./scripts/typecheck.sh` - Run type checking with MyPy
- `./scripts/test.sh` - Run fast test suite (excluding slow/hardware tests)
- `./scripts/check-all.sh` - Run complete quality check pipeline

### Configuration Updates
- **pyproject.toml**: Added balanced MyPy configuration with external library ignores
- **Pre-commit**: Streamlined tooling to avoid conflicts (ruff replaces flake8)
- **Dependencies**: Added all quality tools to dev dependency group

### Code Quality Fixes
- Fixed all import sorting and formatting inconsistencies
- Resolved type annotation issues with proper external library handling
- Added appropriate `noqa` comments for intentional patterns
- Fixed union-attr MyPy issues with runtime `hasattr` checks
- Maintained existing functionality while improving code quality

## Results
- ✅ All tests passing (110 tests)
- ✅ No type checking errors
- ✅ All linting checks pass
- ✅ No medium+ severity security issues
- ✅ Code consistently formatted
- ✅ Pre-commit hooks run automatically on every commit

## Usage
Pre-commit hooks will now automatically enforce code quality on every commit.
Manual quality checks can be run using the provided scripts.

The Voice MCP Server now has production-ready development workflows with
automated quality enforcement while maintaining development velocity.

+++ src/voice_mcp/tools.py
@@ -66,23 +66,23 @@ def _on_hotkey_pressed() -> None:
             # Use real-time transcription with live typing
             stt_handler = get_transcription_handler()
             text_controller = get_text_output_controller()
-            
+
             # Play "on" sound to indicate recording start
             audio_manager = get_audio_manager()
             if audio_manager.is_available:
                 audio_manager.play_on_sound()
-            
+
             # Perform real-time transcription with live typing
             result = stt_handler.transcribe_with_realtime_output(
                 text_output_controller=text_controller,
                 duration=None,  # Use silence-based stopping
                 language=config.stt_language,
             )
-            
+
             # Play "off" sound to indicate recording stop
             if audio_manager.is_available:
                 audio_manager.play_off_sound()
-            
+
             # Log the result
             if result.get("success"):
                 logger.info(
@@ -130,7 +130,7 @@ class VoiceTools:
     def speak(
         text: str,
         voice: str | None = None,
-        rate: int | None = None,
+        rate: float | None = None,
         volume: float | None = None,
     ) -> str:
         """
@@ -226,7 +226,9 @@ class VoiceTools:
             # Handle output mode
             if output_mode != "return" and transcribed_text:
                 text_controller = get_text_output_controller()
-                output_result = text_controller.output_text(transcribed_text, output_mode)
+                output_result = text_controller.output_text(
+                    transcribed_text, output_mode
+                )
 
                 if not output_result["success"]:
                     logger.warning(
